feat: add Pagination component slicing by 5
